+++ Product-Service/src/main/java/org/caesar/productservice/Data/Services/Impl/WishlistProductServiceImpl.java
@@ -49,6 +49,12 @@ public class WishlistProductServiceImpl implements WishlistProductService {
             System.out.println("Sono nella delete");
             Product prod= modelMapper.map(wishListProductDTO.getProductDTO(), Product.class);
             Wishlist wish= modelMapper.map(wishListProductDTO.getWishlistDTO(), Wishlist.class);
+
+            WishlistProduct wishProd= wishlistProductRepository.findByProductAndWishlist(prod, wish);
+
+            if(wishProd == null || wishProd.isOnDeleting())
+                return false;
+
             wishlistProductRepository.deleteByProductAndWishlist(prod, wish);
 
             System.out.println("dopo la delete");
@@ -69,6 +75,9 @@ public class WishlistProductServiceImpl implements WishlistProductService {
             WishListProductDTO wishListProductDTO;
 
             for(WishlistProduct wishlistProduct: wishListProductDTOS){
+                if(wishlistProduct.isOnDeleting())
+                    continue;
+
                 wishListProductDTO = new WishListProductDTO();
                 wishListProductDTO.setWishlistDTO(modelMapper.map(wishlistProduct.getWishlist(), WishlistDTO.class));
                 wishListProductDTO.setProductDTO(modelMapper.map(wishlistProduct.getProduct(), ProductDTO.class));
@@ -89,7 +98,7 @@ public class WishlistProductServiceImpl implements WishlistProductService {
         Wishlist wishlist= modelMapper.map(wishlistDTO, Wishlist.class);
         WishlistProduct wishlistProduct= wishlistProductRepository.findByProductAndWishlist(product, wishlist);
 
-        return wishlistProduct != null;
+        return wishlistProduct != null && !wishlistProduct.isOnDeleting();
     }
 
 
@@ -102,10 +111,11 @@ public class WishlistProductServiceImpl implements WishlistProductService {
             for (WishlistDTO wishlist: wishlists){
                 List<WishlistProduct> wishListProductDTOS = wishlistProductRepository.findAllByWishlist(modelMapper.map(wishlist, Wishlist.class));
 
-                if(wishListProductDTOS.isEmpty())
-                    continue;
-
-                wishListsProduct.addAll(wishListProductDTOS);
+                for(WishlistProduct wishlistProduct: wishListProductDTOS){
+                    if(wishlistProduct.isOnDeleting() && !rollback)
+                        continue;
+                    wishListsProduct.add(wishlistProduct);
+                }
             }
 
             if(wishListsProduct.isEmpty())
@@ -128,7 +138,19 @@ public class WishlistProductServiceImpl implements WishlistProductService {
     @Override  //Rimuove tutti i prodotti della lista desideri passando l'id della stessa
     public boolean deleteAllProductsFromWishlist(WishlistDTO wishlistDTO) {
         try {
-            wishlistProductRepository.deleteAllByWishlist(modelMapper.map(wishlistDTO, Wishlist.class));
+            List<WishlistProduct> wishListProducts = wishlistProductRepository.findAllByWishlist(modelMapper.map(wishlistDTO, Wishlist.class));
+
+            List<WishlistProduct> result= new Vector<>();
+            for(WishlistProduct wishlistProduct: wishListProducts){
+                if(wishlistProduct.isOnDeleting())
+                    continue;
+                result.add(wishlistProduct);
+            }
+
+            if(result.isEmpty())
+                return false;
+
+            wishlistProductRepository.deleteAll(result);
             return true;
         }catch(Exception | Error e) {
             log.debug("Errore nella rimozione dei prodotti dalla lista desideri");
